Move App files into the components directory and merge with calculator.js
