--- src/App.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import { Calculator } from "./components/calculator";
-import './App.css';
-
-export class App extends React.Component {
-  render() {
-    return <Calculator/>;
-  }
-}
